pyproject.toml
[tool.poetry]
name = "projeto-scraper"
version = "0.1.0"
description = "Projeto de coleta de noticias sobre o tema (hacker) na CNN BRASIL, e envio ao email cadastrado."
authors = ["Retr0(kauan)"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.12"
requests = ">=2.32.4,<3.0.0"
beautifulsoup4 = ">=4.13.4,<5.0.0"
python-dotenv = "^1.1.1"




[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
